dispatch actions directly from category facet controller methods
https://coveord.atlassian.net/browse/KIT-1014
